Remove Term() function and put it all in the destructor
We leave Init, so we can give reasonable error codes back to the user.
But BindThread is now baked into Init

Minor, but this also moves the Init options to a struct. I find this
easier to use, since you only have to change the things you want
and can leave the rest default. Instead of having to manually
enter the defaults for all the parameters up to the arg that
you care about.